Add files via upload
This project encompasses data on customer churn from a fictional telecommunication company operating in California during the third quarter. The company provides home phone and internet services to a total of 7043 customers, indicating which customers left, stayed, or subscribed to services. The aim is to predict customer churn using a machine learning model.